feat: make info also return list/dict if required
- also move helper functions to generatedssupersuper so all libNeuroML
  classes have them.
